refactor: move auth logic to lib/auth
